requirements.txt
streamlit==1.19.0
torch==2.0.1
torchvision==0.15.2
Pillow==9.4.0
numpy==1.24.3
opencv-python-headless==4.7.0.72
